source-oracle-batch: Move discovery into a separate file
